ERR: m_Variance and m_MaximumError -> double.

+++ Code/BasicFilters/itkZeroCrossingBasedEdgeDetectionImageFilter.h
@@ -99,10 +99,10 @@ public:
   enum { ImageDimension = TInputImage::ImageDimension };
   
   /** Standard get/set macros for Gaussian filter parameters.  */
-  itkSetVectorMacro(Variance, float, ImageDimension);
-  itkGetVectorMacro(Variance, const float, ImageDimension);
-  itkSetVectorMacro(MaximumError, float, ImageDimension);
-  itkGetVectorMacro(MaximumError, const float, ImageDimension);
+  itkSetVectorMacro(Variance, double, ImageDimension);
+  itkGetVectorMacro(Variance, const double, ImageDimension);
+  itkSetVectorMacro(MaximumError, double, ImageDimension);
+  itkGetVectorMacro(MaximumError, const double, ImageDimension);
 
   /** Get/Set the label values for the ZeroCrossingImageFilter */
   itkGetMacro(BackgroundValue, OutputImagePixelType);
@@ -111,17 +111,17 @@ public:
   itkSetMacro(ForegroundValue, OutputImagePixelType);
   
   /** Set the variance parameter needed by the embedded gaussian filter  */ 
-  void SetVariance(const float v)
+  void SetVariance(const double v)
     {
-      float vArray[ImageDimension];
+      double vArray[ImageDimension];
       for (unsigned int i = 0; i<ImageDimension; ++i) { vArray[i] = v; }
       this->SetVariance(vArray);
     }
   
   /** Set the MaximumError parameter needed by the embedded gaussian filter */
-  void SetMaximumError(const float v)
+  void SetMaximumError(const double v)
     {
-      float vArray[ImageDimension];
+      double vArray[ImageDimension];
       for (unsigned int i = 0; i<ImageDimension; ++i) { vArray[i] = v; }
       this->SetMaximumError(vArray);
     }
@@ -151,11 +151,11 @@ protected:
   
 private:
   /** The variance of the Gaussian Filter used in this filter */
-  float m_Variance[ImageDimension];
+  double m_Variance[ImageDimension];
 
   /** The maximum error of the gaussian blurring kernel in each dimensional
    * direction.  */
-  float m_MaximumError[ImageDimension];
+  double m_MaximumError[ImageDimension];
  
   OutputImagePixelType m_BackgroundValue;
   OutputImagePixelType m_ForegroundValue;
